Add files via upload
This website is built on the basis to help with a university environment enabling a teacher and a student to have a more comprehensive relationship with each other and for better communication system. Unique features accessible to the both the students and teachers, and to quiz themselves over short exams etc and many more. The report shows the absolute aim behind this entire project